src/lambda_worker/requirements.txt
# Lambdaが計算を実行するために必要なライブラリ
skyfield

psycopg2-binary # PostgreSQL接続用ライブラリ
boto3 # AWS SDK
